COMP: Silence warning of mismatched signs.
itkBuildInformationTest.cxx(60): note: see reference to function template instantiation
'testing::AssertionResult testing::internal::CmpHelperGT<unsigned int,int>(const char *,const char *,const T1 &,const T2 &)'
being compiled with
          [
              T1=unsigned int,
              T2=int
          ]

+++ Modules/Core/Common/test/itkBuildInformationTest.cxx
@@ -57,7 +57,7 @@ TEST(ITKBuildInformation, InformationFeatures  ) {
 
   EXPECT_EQ(itk::BuildInformation::GetValue("PROJECT_URL"),"http://www.itk.org");
 
-  EXPECT_GT( itk::BuildInformation::GetAllKeys().size(), 5 );
+  EXPECT_GT( itk::BuildInformation::GetAllKeys().size(), static_cast<size_t>(5) );
 
   for(const auto keyEntry: itk::BuildInformation::GetAllKeys() )
   {
